LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/back-forward-cache/service-worker-clients-matchall.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Invalid URL

Harness Error (TIMEOUT), message = null

TIMEOUT Clients.matchAll() should not list pages in BFCache Test timed out
